feat: 외부 클릭 시 callback 구현 훅 추가
